+++ lib/view_component/scoped_styles/railtie.rb
@@ -6,6 +6,10 @@ require "rails/railtie"
 module ViewComponent
   module ScopedStyles
     class Railtie < Rails::Railtie
+      initializer "view_component.scoped_styles.component_generator_hook" do
+        require_relative "component_generator_hook"
+      end
+
       config.after_initialize do
         next unless Rails.env.development?